SURF - widthSample is now a double - Done to make tuning easier since it is multiplied by the scale (which is a double) and now finer grained values can be selected - Improved documentation

+++ main/boofcv-feature/src/main/java/boofcv/alg/feature/describe/DescribePointSurfMod.java
@@ -69,7 +69,7 @@ public class DescribePointSurfMod<II extends ImageGray<II>> extends DescribePoin
 	 * approximation will be used.  True is recommended.
 	 */
 	public DescribePointSurfMod( int widthLargeGrid, int widthSubRegion,
-								 int widthSample, int overLap,
+								 double widthSample, int overLap,
 								 double sigmaLargeGrid, double sigmaSubRegion,
 								 boolean useHaar, Class<II> imageType ) {
 		super(widthLargeGrid, widthSubRegion, widthSample, 1, useHaar, imageType);
